requirements.txt
aiofiles>=23.0.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
Django==5.1.6
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
google-api-core>=2.24.0
google-auth>=2.38.0
google-cloud-vision>=3.10.0
googleapis-common-protos>=1.69.0
grpcio>=1.71.0
grpcio-status==1.71.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.11
lxml==5.3.1
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
pymilvus==2.4.10
milvus-lite==2.4.11
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0

nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.21.0
openai==1.65.2
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20240706
pi_heif==0.21.0
pikepdf==9.5.2
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pypandoc==1.15
pyparsing==3.2.1
pypdf==5.3.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
timm==1.0.15
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
unstructured==0.17.0
unstructured-client==0.31.1
unstructured-inference==0.8.9
unstructured.pytesseract==0.3.15
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.2
xlwt==1.3.0
yarl==1.18.3
zstandard==0.23.0
json5==0.9.25
